+++ lib/liquid_glass_helper.dart
@@ -20,6 +20,8 @@ class LiquidGlassHelper {
         'isLiquidGlassSupported',
       );
       return supported;
+    } on MissingPluginException {
+      return false;
     } on PlatformException {
       return false;
     }